Update CPU-Info to support M4 Pro/Max and uplevel enum to string (#57)
Fixes the segfault when CPU Uarch exceeds the max size of `magic_enum`
and adds support for M4 to cpuinfo.
@sjw36 and I discussed offline and we are not sure uarch is useful for
the M series, but decided to continue using `uarch` for now until we
find a compelling reason not to. There are some interesting limitations
reading uarch alone with Apple, e.g.
https://github.com/llvm/llvm-project/pull/95478#issuecomment-2166923128